kafkaIntegrationService/requirements.txt
confluent_kafka==2.6.1
PyYAML==6.0.2
Requests==2.32.3